fix: simplify tests
Get pids based on the name of the project/command. This means
we need to parse the state directory to get the command to then search
for the pid. This actually works out quite nice, since we then asserts
that the state file exists when getting the pids, not needing to
remember to assert it every time in each test